Move Integration tests to the IntegrationTests project
