+++ src/Helm.Tests/Helm.Tests.csproj
@@ -15,5 +15,4 @@
   <ItemGroup>
     <ProjectReference Include="..\Helm\Helm.csproj" />
   </ItemGroup>
-
 </Project>
